cache second mapping
